finished day 15 part 1

+++ template/src/main.rs
@@ -19,14 +19,14 @@ fn parse(s: &str) -> IResult<&str, Vec<Vec<u32>>> {
     separated_list1(newline, separated_list1(space1, u32))(s)
 }
 
-fn process(s: &str) -> u32 {
+fn process(s: &str) -> usize {
     let (_, input) = parse(s).unwrap();
     println!("{:?}", input);
 
     todo!()
 }
 
-fn process2(s: &str) -> u32 {
+fn process2(s: &str) -> usize {
     let (_, input) = parse(s).unwrap();
     println!("{:?}", input);
